#2893: consolidate d3 animation with 'react-faux-dom'
