+++ src/jsx/import/content/home-page.jsx
@@ -15,7 +15,6 @@ import React, { Component } from 'react';
 import { NavLink } from 'react-router-dom';
 import { setLayout } from '../redux/action/page.jsx';
 import Animate from '../animation/animate.jsx';
-import collision from '../animation/node/collision.js';
 
 class HomePage extends Component {
     componentWillMount() {
@@ -26,7 +25,7 @@ class HomePage extends Component {
         return (
             <div className='main-full-span home'>
                 <h1>Welcome!</h1>
-                <Animate node={collision} />
+                <Animate />
                 <NavLink
                     to='session'
                     activeClassName='active'
